requirements.txt
seleniumbase
python-dotenv
mysql-connector-python